Update to common source code; does not affect open-vm-tools.
Allow an asyncWebSocket to connect through an existing socket descriptor

Because Windows does not allow unprivileged processes to specify
DSCP (formerly ToS) values on TCP traffic, in order to support
DSCP for Blast WebSockets we must obtain a preconfigured socket
descriptor from a privileged process and then build the WebSocket
connection over that socket.

This changeset extends the asyncWebSocket API by adding a Connect
function that can accept and use an existing socket descriptor in
addition to the usual collection of WebSocket Connect parameters.

Because the asyncWebSocket is built over an asyncTCPSocket, this
change ripples down into the asyncTCPSocket API which also gets a
new Connect variant with a socket descriptor parameter.

To avoid duplicating existing logic, the old Connect variants
are modified to do their work by invoking the new API with
a distinguished socket descriptor value of -1.        This value
indicates that no existing socket descriptor is provided and
that a new socket should be created and used for the connection.

In this changeset, passing in an existing socket is supported
only on the Windows platform.  The feature is not required on
other platforms, where unprivileged processes are permitted to
define DSCP values for their connections.  Attempting to create
a connection over an existing socket on other platforms will
result in a Connect failure.