Add support for NVMe disks; IDE and SATA disk devices in 3.x Linux kernels
Support has been added for virtual disks attached to an NVMe
controller.  The device "label" will provide the "nvme<n>" for the
controller.  The disk unit is provided in the device "nsid" file.
An earlier implementation (Linux 2.6 kernels) provided the unit
number as the final digit group in the device node name.

The Linux 3.x kernel has some differences in the layout of IDE and
SATA disk devices in the sysfs filesystem.  The differences from the
sysfs in a 2.6 Linux kernel are:
 - The IDE "class" file and SATA "class" and "label" files are in a
   different directory.
 - IDE controller and SATA device nodes begin with "ata" where "host"
   was used in earlier kernel sysfs filesystems.

GuestInfoGetDevClass() has been modfied to check in two possible
locations for the "class" file.   The directory located will also
contain the "label" file for SCSI, SAS, and SATA disk devices.

GuestInfoGetIdeSataDev() has been modified to search for either
ata<n> or host<n> directories in the sysfs tree for the disk device.