Fix the deployPkg status files when privateTmp is enabled
Currently, deployPkg status file path is hardcoded to /tmp.  When
privateTmp is enabled in the guest VM, /tmp is hidden from the rest
of the system and is wiped on reboot.  So need use different path to
save the status file.

This change is changing deployPkg status file path from /tmp to /var/log