Variadic LOG macros and fewer trailing newlines
It's (long past) time we started using variadic LOG macros
and stopped requiring a newline at the end of every format
string.  A previous removed the newline requirement recently.

-   LOG(N, ("format string\n", arg1, arg2))
+   LOG(N, "format string", arg1, arg2)

The important parts of this change are buried in macro madness.
The key bit is the LOG_BYNAME macro, which now can be written
to be variadic.

To support both styles simultaneously, this change adds a macro
LOGLEVEL_VARIADIC which switches the definition of LOG_BYNAME to
variadic (e.g. remove extra parens).

Following this change, we can convert files to the variadic version
and set LOGLEVEL_VARIADIC.