Fix a minor issue reported by Codacy code scanner.
One of the code quality checkers (Codacy) reported the following
issue for a bunch of files bundled in 'open-vm-tools'.

"""
Fix 'variable has been re-assigned before it's value has been used'
"""

Checked all applicable files and removed the variable initialization
where it's really not necessary.

This is really not a major issue but it's no harm fixing.