Fix 'variable scope can be reduced' warnings.
One of the code quality checkers (Codacy) reported the following
issue for a bunch of files bundled in 'open-vm-tools'.

"""
Fix 'The scope of the variable 'var' can be reduced.'
"""

Checked all applicable files and removed the variable initialization
where it's really not necessary.

This is really not a major issue but it's no harm fixing.