Fix misc. warnings related to strncpy usage.
For few files bundled in open-vm-tools package, one of the code
scanners reported the following warning.

" Easily used incorrectly; doesn't always \0-terminate or check for
invalid pointers"

Checked the files and modified the code to pass 'sizeof dest - 1'
as the third argument to avoid any warnings.

In one of the VGAUTH library functions, added a check for the input buffer
size and return a valid error.