Address some issues reported by the Coverity scan of open-vm-tools.
Fix or annotate issues reported in the Coverity scan of open-vm-tools,
as follows:

(1) Fix NULL pointer de-reference in failure case of
CopyStringListEntry.

(2) Add NULL pointer checks before dereferences in
VixMsg_DeObfuscateNamePassword.

(3) Annotate a false positive in VixMsg_EncodeString.

(4) Annotate and add an ASSERT to a false positive in
StrUtil_GetLongestLineLength.