hostinfoPosix.c: avoid segfaulting on invalid lsb_release
If lsb_release is broken it might return an empty string in stdout which
will later on crash the program. After the check for NULL (totally
invalid answer) also check for strlen being zero.

In those cases not only the crash is avoided, in addition also there
would have been no useful data in that string and it is better to use
the fallback of HostinfoReadDistroFile instead.

Fixes: #390

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>