hostinfoPosix.c: check for valid pointer in HostinfoLsbRemoveQuotes
HostinfoLsbRemoveQuotes requires the caller to pass non-NULL pointer,
otherwise the following free calls in the caller would break anyway.

Never the less a check to !=NULL is defensive programming and this isn't
a fast path.

Furthermore a char* could also be empty which would segfault when accessing
lsbStart[0]. To stay on the safe side in case more callers to
HostinfoLsbRemoveQuotes are added in the future at least check the string
length before accessing lsbStart[0].

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>