Avoid vmtoolsd crash in HostInfo.
The guest identification code causes vmtoolsd to crash in certain
versions of some distros.  The crash is caused by recent changes to
the lsb_release command.  Previously, if the command existed, all
of its options worked.  Now, some of the options no longer exist.

Change the code to check for an lsb_release failure whenever it is
invoked.

Fix for: https://github.com/vmware/open-vm-tools/issues/390