Add arm64 support in open-vm-tools
This adds the required files and fixes to be able to build
the open-vm-tools for arm64.

One major change is the update of autoconf from 2.61 to 2.69,
with automake.  The autoconf update forced to add the libtool
include path to properly run autoreconf on the tools.  I don't
know if these AUTOCONF/AUTORECONF variables are used anywhere
else.

The open-vm-tools have been tested on debian 10 and fedora 31
VMs (thanks to Andreas Scherr).

Also fixed Hostinfo_TouchVirtualPC() which was incorrectly
testing if vm_x86_64 was defined.  It is always defined, to
either to 0 or 1.