Changes to common header files. Make ALIGNED() functional for 64-bit Windows
The current implementation of ALIGNED in vm_basic_types.h only
works for compilers that are, or pretend to be, GCC.  However,
it is reasonable to conclude that data specifically aligned
for one OS / compiler should be aligned the same way when built
with other tools or for OSes.

This change updates the ALIGNED macro to correctly function when
building on 64-bit Windows.  If not using GCC-alike, or Windows,
the macro continues to expand to nothing.

The MSVC __declspec construct accepts a limited syntax, basically
requiring an integer literal (see pciPassthruInfo_public.h).  It
must be a prefix operation.

The GCC __attribute__ is more expressive and can be postfix or
prefix, so it is moved to be prefix.

The bulk of the change moves the use of ALIGNED to be prefix in
all places that compile for the Workstation product on Windows.