Handle LSI Logic Parallel and BusLogic SCSI on PCI bus 0.
When a LSI Logic Parallel or BusLogic SCSI controller appears on PCI
bus 0, the disk device controller name is not present; the "label" file
is missing.  Adding logic to check if disk device is attached to a
SCSI controller on bus 0 (0000:00:10.0) if a "label" file cannot be
found.  If that is the case, setting the device name to "scsi0.n"
where "n" is the device unit number previously determined.