Handle Photon's "/dev/root" mount point when doing disk device mapping
Photon EFI boot VMs have the root disc mounted long before the various
PCI controllers are initialized and the attached devices enumerated.
When looking through the /proc/mounts, the root filesystem mounted
at '/' is displayed as if on device /dev/root/  But there is no "root"
device in /dev, nor in the PCI device tree at /sys/class/block.  Later
Photon builds fabricate a block device at "/dev/root" but again nothing
in the /sys/class/block PCI tree.

For this situation, it is necessary to use the pseudo device's major and
minor number to access the PCI device tree through /sys/dev/block.
The major and minor device number can be extracted from the contents of
/proc/self/mountinfo.

19 0 8:2 / / rw,relatime shared:1 - ext4 /dev/root rw

The 3rd field is the major:minor number and the 9th or 10th, depending
on Linux release, is the pseudo device in this case.