Replace VixTools_ConfigGetBoolean with VMTools_ConfigGetBoolean
Initially came up with a utility function named VixTools_ConfigGetBoolean
to retrieve the boolean value from the config dictionary.  This function
was designed as local/specific to VIX plugin.  Later, VMTools_ConfigGetBoolean
function was designed with the same functionality and is being used everywhere
in the Tools repository.

This changeset gets rid of VixTools_ConfigGetBoolean and replaces
all those occurrences with VMTools_ConfigGetBoolean.