Common source file changes not directly applicable to open-vm-tools.
Fix procMgr library for Windows.

- For windows, to retrieve the command line for a target process,
the procMgr library creates a remote thread in the target process
and executes GetCommandLine function.  This approach is OK. But for
few applications that are not built with proper flags, creating a
remote thread may crash the application.  In order to avoid any
issues, it was decided not use the 'remote thread' approach by DEFAULT.

- Refactored the procMgr APIs to take an input argument from the
user if the 'remote thread' approach needs to be used.

- Refactored the procMgr APIs to conditionally use the WMI for
retrieving the command line for a target process.

- New options are provided to the user if the 'remote thread'
and 'WMI' approaches need to be forced.  Modified the VIX tools plugin
to read and honor the new flags from tools.conf

The following are the new flags under guestoperations group.
useRemoteThreadForProcessCommandLine
useWMIForProcessCommandLine

The default value for the above new flags is false. User can set them
to true if needed.

- Re-factored some code in appInfo plugin that calls procMgr library.
The appinfo plugin doesn't really use the commandline.  So, modified the
appinfo plugin to just use the API that doesn't use the complicated
approaches for the listing down the processes.