Fix a Coverity-reported overrun.
A Coverity scan of open-vm-tools reports a buffer overrun in
Escape_Unescape.  The problem is that Escape_Unescape uses
sizeof('\0') to specify the size of a buffer that consists of
a single character in the variable nulByte (previously named
nullbyte).  However, character literals in C are ints, so
sizeof('\0') is equivalent to sizeof int rather than sizeof char.
Use "sizeof nulByte" instead.