Make Backdoor fallback temporary.
When RpcOut falls to Backdoor, it stays with Backdoor
permanently for the life of vmtoolsd service.  It is a
long standing bug in the reset handling code.  Typically,
channel type is not changed during reset.  Our reset
handling code can either keep the channel type same or
switch it from vsocket to Backdoor, but it can't do other
way.  Though it is supposed to switch to vsocket on reset
caused by events like vmtoolsd being restarted or VMX
breaking the channel for some VM management operation.
With this change when we start the channel, we always
try vsocket first unless Backdoor is enforced by the
caller.

Using Backdoor for too long is not desirable because
privileged RPCs can't be used on such channel.  So, we
need to retry switching the channel back to vsocket
periodically.  We don't want to try vsocket on every
RpcChannel_Send call because that adds to overhead and
increases the latency of RpcChannel_Send due to connection
timeouts.  So, we retry vsocket with a backoff delay
between 2sec-5min.

As some RpcChannel callers intend to use Backdoor channel
we need to differentiate between such usage from the
callers that create vsocket channel and fallback to
Backdoor.  Therefore, introduced a concept of mutable
channel.  The vsocket channel is mutable as it can fallback
to Backdoor and restore vsocket.  However, if a caller
creates Backdoor channel, it will not be mutable and
stay with Backdoor for its lifetime.

As vmxLogger frequently connects and disconnects the
channel for every log message and does not use any
privileged RPC, so make it use Backdoor channel
permanently to avoid frequent vsocket connections.

Additionally, removed the redundant 'stopRpcOut' interface
and renamed 'onStartErr' to 'destroy'.