Use random poll-interval for appInfo on channel reset.
In few workflows like instant clone, when a large number of VMs are
cloned at the same time, if the appinfo plugin runs at the same time
in all the VMs, the underlying ESXi may encounter heavy load.  To
avoid these situations, a random poll interval should be used
for the appinfo whenever applicable workflows are detected.

Detecting a 'rpc channel reset' is a simple approach to detect.

In this changeset, add the following changes:

- Added a new callback function for the 'rpc channel reset'.  If
the rpc channel is reset, a new random poll interval is calculated
and poll timer is adjusted accordingly.  If the existing appinfo
poll interval is greater than the minimum interval of 30 seconds,
random interval will be generated between 30 and appinfopollinterval.
If the existing poll interval is less than the minimum 30 seconds
time, then random interval will not be changed.

- Code refactoring for few functions.
- Changed one global variable as static.
- Added few debug/info log messages.