[AppInfo] Tweak the gather loop only for a real config reload.
The poll loop for the appInfo is being tweaked (destroyed and recreated)
for every conf reload even when nothing related to appinfo changed.
This may cause few scenarios where the 'application information'
will never be collected inside the guest.  Fixed the code, to
tweak the loop only when there is a real appinfo related config change
in the tools.conf file.