Improve logging on signature verification
If xmlsec and vgauth are built inconsistently, xmlSecSize
can differ between the two, which results in a key datastructure
being a different size, and vgauth ends up looking in the
wrong place in memory for the signature check status.

Log the status, so that if its not one of the two
expected values, we have some ideas of what went wrong.