Hgfs Linux/OSX Server: fix write-only shares access check creating new files
Linux and OS X Hgfs server has an incorrect failure status check when a
user has a write-only share enabled.  In a write-only share any failure
is mapped to EACESS unless the error is ENOENT when a new file is to be
created.  The error check currently fails all errors when it should
only allow creation of new files i.e., ENOENT with flags specifying
O_CREAT.

The check should be
   if (status == EACCES) goto exit