lib/file: Fix a memory leak in FileLock_Lock.
Valgrind complained about a minor leak within FileLock_Lock when running
test-vmx Firmware.Efi.QuickBoot .  FileLockScanner builds a list of "active
locks" and traverses that list repeatedly, but whenever it decides to remove
an entry from the list, it would simply leak it.

This change ensures that the list item is not leaked by Posix_Free()ing it
once it is no longer needed.