Improve Linux guest identification function documentation.
Make it clear that the "short name" of a Linux distro has a default
value and that the default value should not be overwritten unless
VMware does this.

If someone cheats and sends down an unsupported "short name", the
guestMapper will protect the software stack, mapping the unsupported
"short name" to that of the guestOS config found in the VMX file.
We know that is OK since the VM couldn't power on and run the guest
unless it was valid.

All older, supported ESXi have updates that contain the guestMapper.

If the guestMapper is not in place - an older, unpatched release -
the software stack can become confused causing problems and crashes.

Change augments https://github.com/vmware/open-vm-tools/pull/431