hostinfoPosix: remove NOT_IMPLEMENTED() when NO_IOPL
When running "vmtoolsd -b pidfile" on Linux arm64, we hit this error:

   [error] [vmsvc] NOT_IMPLEMENTED hostinfoPosix.c:2526

Linux arm64 does not implement iopl() and the Hostinfo_ResetProcessState()
is raising a NOT_IMPLEMENTED() because of that.  However if there's no iopl(),
there is no reason to drop IO privileges, so we can just skip that code.

This change also restores usage of getloadavg() for non Android arm64
platforms: ESX vmx and tools for Linux arm64.