GOSC event doesn't report at once after customization process timer is reached
When process timer is reached, the command process is killed and function
ProcessRead is called to read all the output till EOF.  But EOF can't be
reached immediately.  The reason is pre-customization script is launched
by perl script, and killing perl script process doesn't kill the
pre-customization script process.
This code update includes:
1. Do not read stdout/stderr when gosc command process exits abnormally
   to avoid the EOF blocking.
2. Add READSTATUS_PENDING_TO_EOF in enum ReadStatus to avoid the confusion
   with READSTATUS_PENDING.
3. Close the write ends of pipes (stdout[1]/stderr[1]) for child command
   process before it exists.
4. In processPosix.c, the write ends of pipes have been closed in line 180,
   181, so the read ends should be closed in line 254, 255.
5. Add explicit note for the beginning and the end of the perl script log.