[HGFS Server] Support oplock inside HGFS server for Windows host
The motivation for this change is to cache file/directory related information
in host side, for example the file/directory attributes.  To make the cache
correct, we will use the oplock(Windows)/lease(Linux) to monitor the
file/directory change event.  When the file/directory changes, the item in
cache will be invalidated.

In this change, two new functions are defined:
- HgfsOplockMonitorFileChange
   This function is used to monitor the change event for a file/directory,
   and the callback will be called if file/directory is changed.
- HgfsOplockUnmonitorFileChange
   This function is used to cancel the change event monitoring.

This patch only implements the oplock support for Windows host, the support
for Linux will be delivered in another patch.