[AppInfo] Tweak the gather loop only for a real config reload.
The poll loop for appInfo is being tweaked (destroyed and recreated)
for every tools.conf reload, even when nothing related to appinfo changed.
This may cause a few scenarios where the 'application information'
will never be collected inside the guest.  Tweak the loop only when there
is a real appinfo related config change in the tools.conf file.