LOG: Keep the LOG_LEVEL entries sorted
This is the first in a series of changes to the Log Facility that will add
the ability to have domain specific filtering (e.g. foobar domain can have
log calls which are filtered by domain (name) and level).  This will make
LOG like abilities usable in all build types, even a release build.

The LOG_LEVEL entries are difficult to find since they are in a jumbled
order.  The existing grouping isn't useful as many of the modules are shared.
Create one big sorted list.