build: fix propagation of libtirpc flags
open-vm-tools already detects libtripc in configure:
  checking for libtirpc (via pkg-config)... yes
  configure: tirpc is needed: yes
  configure: building with libtirpc
And build uses it most of the time:
  ... -DUSE_TIRPC -I/usr/include/tirpc ...
  ... -ltirpc ...

Yet with glibc2.32 finally dropping --enable-obsolete-rpc that finally
removes /usr/include/rpc/rpc.h and such and only leaves the proper
/usr/include/tirpc/rpc/rpc.h around.
This causes several build issues as the libtripc flags are in XDR_
variables which are not present in all places where XDR related headers
are includes.

Add some more of these flags in the places that got flagged by build
fails to fix compilation with glibc >=2.32.

Fixes: #486

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>