[GlobalConf] Support for vmusr service to load Global Configuration.
Currently, the vmsvc service periodically downloads the Global Configuration
from the GuestStore and applies the downloaded configuration.  'vmusr'
service doesn't have the support.  In this changeset, made the changes
for both vmusr and vmsvc services to periodically check and apply
the changes.

To keep thing simple and easy, the new approach is check the
modification time every time the regular tools conf is read and if
changes are detected, the configuration is applied.  With this new approach,
signalling mechanism is not needed and that code is removed.

vmsvc:
* Starts a background thread which periodically fetches the global
configuration from the GuestStore.

vmusr:
* No background thread is created. But checks and loads the configuration
everytime the regular tools.conf is checked.