Change the default Unicode error errno.
In times long past, EINVAL was chosen for the errno returned when a
Unicode problem was detected.  This overloads EINVAL, particularly when
using POSIX system calls.  Unfortunately, the overloading makes it very
difficult to recover from some errors as there is no way to know the
difference between a "real" EINVAL and a Unicode problem EINVAL.

Over 10 years of data shows that we *RARELY* encounter Unicode errors.
Inspecting the source base shows the code segments that do check for
errno after an error either:

1) Post the errno value and die.
2) Have switch handle cases but also have a default catcher.
3) Have an if/else waterfall with a final else.

Another observation is that Unicode routines are not involved with
anything that would return a math related errno (e.g. ERANGE, EDOM).

Changing the value of UNICODE_CONVERSION_ERRNO to ERANGE and avoid
any overloading.