Log file name becomes invalid after a rotation
This is because the accounting is done unsigned but the printf used (in
multiple places) was "%d".  Fix this by using "%u".

As documented in the function header, the wrap around case was not handled
properly, so this was fixed as well.  If the maximum rotation number hits
MAX_UINT32, all of the files are renamed to pack the files as if this was
the beginning of a rotation sequence.