Add a switch to the vm-support script to transfer the support bundle to the hypervisor
The vm-support script unconditionally transferred the support bundle to
to the VMware hypervisor, thereby filling vmware.log.
This fix changes the default behavior of the Linux and Windows scripts
to *not* transfer the logs to the host.

Added an option '-x' to both scripts to let the user elect to
transfer the guest support bundle to the host.