resolution: Fix kms autodetection
Currently, the elf binary of the xorg driver is loaded to check for a
string that was put in the .modinfo section in the driver.  Unfortunately
there are two problems with this approach:

  1) Distros now ship without xorg, so the xorg .so doesn't exist and
     there's nothing to check.
  2) Distros (e.g. Fedora) do heavy optimizations and remove the .modinfo
     section from the .so, so the string cannot be found even though the
     driver exists.

To fix both, stop depending on being able to parse the elf binary of the
xorg driver.  Instead, let the plugin check for the existence of the drm
driver with a sufficiently high version, and if it exists, use kms.
This removes the dependency on X for kms.  Also increase the version of
vmwgfx required to ensure atomic mode-setting plus relevant bug fixes (in
the kernel since 2017) are available.