Implement log redirection for the deployPkg plugin.
On Linux, /var/log/vmware-imc/toolsDeployPkg.log is the default deployPkgs
log file.  This change allows users to redirect that log to the vmware.log
on the host or to another file located on the guest VM.  Redirection is
controlled by settings in the "logging" section of the tools.conf file.

See the examples in the sample tools.conf file provided in the distribution.