Linux guest identification: Support multiple architectures.
Allow machine architecture information to be present in the
"short string" and the guestOS string.  This is done via a prefix,
"<arch>-" (e.g. arm-ubuntu-64, riscv-debian11-64, arm-windows10-64)
with the X86 architecture being implied (no explicit "<arch>-").

Add the recognition of the Arm machine architecture to the Linux
guest identification code.