Asyncsocket: Add comments about AsyncSocket_SetErrorFn usage.
Document that AsyncSocket_SetErrorFn must be called before an internal
asyncsocket callback can fire and trigger a call to the error handler.
The error handler must be set immediately after the asyncsocket is created,
either from the poll thread, which requires no additional synchronization,
or while holding the asyncsocket lock, which is passed via pollParams.