Log messages when the network interface limit is hit.
Whenever the maximum NIC limit is reached, log a message to both the
VM guest.log file on the host and the vmsvc.log file inside the guest.

Moved the logging api added for powerOps plugin to vmtoolslib so
that it is available for all plugins.  Modified the powerOps code
accordingly.