Fix an ASSERT in bora/lib/misc/timeutil.c.
Fix a problem with TimeUtil_NtTimeToUnixTime on Arm that was encountered
when running tools tests on Apple silicon.  The problem was the routine
assumed that a variable was 32-bits if VM_X86_64 was not defined.  This
may have been true in the past, but it is no longer true now that the code
is also built for 64-bit Arm.