Remove unwanted "volatile" from static variable used with g_once_init_enter()
The glib api documentation for g_once_init_enter() specifically states
that "volatile" should NOT be used with the address passed as the
argument.  Recent compilers (GCC-11 & clang 11) and recent versions of
glib-2 will result in a warning that the "volatile" qualifier has been
dropped if it has been used.

Remove the unneeded and unwanted "volatile" qualifier from the definition
of "inited" in pollGtk.c.

Fixes: https://github.com/vmware/open-vm-tools/issues/509