Customization: Retry the Linux reboot if telinit is a soft link to systemctl.
Issues have been reported on some newer versions of Linux where the VM
failed to reboot at the end of a traditional customization.  The command
'/sbin/telinit 6' exited abnormally due to SIGTERM sent by systemd and
where telinit is a symlink to systemctl.

This fix checks if telinit is a soft link to systemctl and if yes, retries
a system reboot until the telinit command succeeds or the reboot literally
happens.