Update open-vm-tools to build with either Fuse 3 or Fuse 2
Vendors are requesting that open-vm-tools can be built with either
Fuse 2 or Fuse 3.  While both Fuse 2 and Fuse 3 runtime can be
installed on a Linux system, vendors would prefer to switch from
Fuse 2 to Fuse 3 at the same time for all products to be available
with the base OS.

Updating the configure.ac file to check for the presence of the Fuse 3
or Fuse 2 development packages in the build environment.  Providing
configure options to allow users to control the version of Fuse to be
used.
 --without-fuse       - vmblock-fuse and vmhgfs-fuse will be disabled.
 --with-fuse=fuse3|3  - use Fuse 3.x
 --with-fuse=fuse|2   - use Fuse 2.x
 --with-fuse=auto     - check for Fuse 3 or Fuse 2 availability; disable
                        vmblock-fuse and vmhgfs-fuse if unavailable.
 --with-fuse          - implicit "yes".
 --with-fuse=yes      - check for Fuse 3 or Fuse 2 availability; disable
                        vmblock-fuse and vmhgfs-fuse if unavailable.

Pull request: https://github.com/vmware/open-vm-tools/pull/544
Fixes issue:  https://github.com/vmware/open-vm-tools/issues/314

The vmblock-fuse code is also used by WorkStation.  Configure defines
are not available in internal builds.  Reworked preprocessor tests to
use FUSE_MAJOR_VERSION from the fuse headers to determine API to be
used during compilation.