AppVisibility: Skip terminated/stopped containers.
Added code to skip terminated/stopped containers.  This is done
by using TaskStub to retrieve the pid for the running container.
If the pid cannot be retrieved, then that container is skipped.

The previously suppressed "default" namespace has been restored.

Fixed another minor issue with calculating the number of containers
that were added.

Note: This changeset introduces a dependency on Tasks.proto, and
Tasks.proto depends on a few other proto files like mount, metrics,
descriptor and task.  Modified the makefile to generate the necessary
C++ files from the proto files.

Additional notes:
 - .proto -> generates .pb.h, .pb.c and .grpc.pb.cc files.
 - For compiling and building the library (.so file), the .grpc.pb.{h,cc}
   and .pb.{h,cc} files are needed.
 - The protoc compiler generates .pb.h and .pb.c files when the --cpp_out
   option is specified.
 - The protoc compiler generates .grpc.pb.* when the --grpc_out option is
   specified.