Make Linux perl based customization work with the cloud-init workflow.
To resolve issues seen where users want to set a vm's networking
and apply cloud-init userdata together before the vm is booted, the
deployPkg plugin has been modified to wait for cloud-init
execution to finish.  This allows cloud-init to finish execution
completely before the customization process triggers a reboot
of the guest.

This change is solely in the deployPkg plugin side, so a user can get
this change by upgrading their open-vm-tools in the guest/template.