Add explanatory comments for Coverity false positives.
Add comments for two memory leak false positives reported by a
Coverity scan of open-vm-tools.   Don't annotate since the
annotations can't be made specified to the leaked variable, so
that if an actual leak were introduced in the future an annotation
would cause it to be reported as a false positive.

For the same reason, replace a leaked storage annotation added
previously with an explanatory comment.