Avoid a vmtoolsd service error message if not on a VMware hypervisor.
When open-vm-tools comes preinstalled in a base Linux release, the
vmtoolsd services are started automatically at system start
and desktop login.  If running on physical hardware or in a non-VMware
hypervisor, the services will emit an error message to the systemd's
logging service before stopping.

This change removes the unwanted error message.