Enhance the guest identification code for Linux
The Linux guest identification code is enhanced to return two
additional fields (when a distro makes them available):

1) The VERSION field

This field sometimes contains additional information not
found in other fields. For instance, on SLES, this provides
the patch level information.

2) The CPE_NAME field

This is the NIST Common Platform Enumeration Specification
string. If present, this may provide information in a
standardized form.