[TimeInfo] Introduce TimeInfo in TimeSync plugin
This change lays a foundation for upcoming changes to
support TimeInfo feature in open-vm-tools. TimeInfo feature
is introduced as part of TimeSync plugin and can be
used to query, set, subscribe, and receive updates for
time-related information from the host when guest is
using precisionclock to consume time from the host.

This change simply adds a new file and basic init/shutdown
routines which are called as part of TimeSync plugin
load/unload. The change also introduces a config option
to enable/disable this feature (default is off). The
feature is Linux-only for now.

Upcoming changes will add support for subscribing and
receiving TimeInfo updates.